adding the config file to spawn-agent, refactoring for common usage + typo
